+++ gpaw/calculator.py
@@ -84,9 +84,7 @@ class GPAW(PAW, Calculator):
         'convergence': {'energy': 0.0005,  # eV / electron
                         'density': 1.0e-4,
                         'eigenstates': 4.0e-8,  # eV^2
-                        'bands': {'occupied': True,
-                                  'energy': None,
-                                  'number': None},
+                        'bands': 'occupied',
                         'forces': np.inf},  # eV / Ang
         'dtype': None,  # deprecated
         'width': None,  # deprecated
@@ -737,14 +735,11 @@ class GPAW(PAW, Calculator):
         self.log('Number of atomic orbitals:', self.wfs.setups.nao)
         self.log('Number of bands in calculation:', self.wfs.bd.nbands)
         self.log('Number of valence electrons:', self.wfs.nvalence)
-        self.log('Bands to converge: ', end='')
-        p = par.convergence.get('bands', {'occupied': True})
-        if p == 'occupied':
-            p = {'occupied': True}
-        elif p == 'all':
-            {'number': 'all'}
-        elif isinstance(p, int):
-            ...
+
+        n = par.convergence.get('bands', 'occupied')
+        if isinstance(n, int) and n < 0:
+            n += self.wfs.bd.nbands
+        self.log('Bands to converge:', n)
 
         self.log(flush=True)
 
@@ -871,8 +866,7 @@ class GPAW(PAW, Calculator):
         nbands_converge = self.parameters.convergence.get('bands', 'occupied')
         if nbands_converge == 'all':
             nbands_converge = nbands
-        elif nbands_converge != 'occupied':
-            assert isinstance(nbands_converge, int)
+        elif isinstance(nbands_converge, int):
             if nbands_converge < 0:
                 nbands_converge += nbands
         eigensolver = get_eigensolver(self.parameters.eigensolver, mode,