+++ gpaw/eigensolvers/cg.py
@@ -62,7 +62,7 @@ class CG(Eigensolver):
                              'over %d band groups.' % wfs.bd.comm.size)
         Eigensolver.initialize(self, wfs)
 
-    def iterate_one_k_point(self, ham, wfs, kpt):
+    def iterate_one_k_point(self, ham, wfs, kpt, weights):
         """Do conjugate gradient iterations for the k-point"""
         self.timer.start('CG')
 
@@ -222,13 +222,7 @@ class CG(Eigensolver):
                         break
                     error = error_new
 
-            if kpt.f_n is None:
-                weight = 1.0
-            else:
-                weight = kpt.f_n[n]
-            if self.nbands_converge != 'occupied':
-                weight = kpt.weight * float(n < self.nbands_converge)
-            total_error += weight * error
+            total_error += weights[n] * error
             # if nit == 3:
             #   print >> self.f, "cg:iters", n, nit+1
         if self.tw_coeff:  # undo the scaling for calculating energies